pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "safestream"
version = "0.1.0"
description = "A Real-Time Moderated Live-Chat Simulator"
authors = [
    {name = "SafeStream Team", email = "team@safestream.dev"}
]
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-dotenv",
    "black",
    "ruff",
    "pytest",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-multipart",
    "bcrypt<4.1.0",  # Pin to compatible version
    # TODO(stage-7): "streamlit",  # Moderator dashboard
    # TODO(stage-5): "sqlalchemy",  # Database ORM
    # TODO(stage-8): "locust",  # Load testing
    # TODO(stage-5): "pydantic",  # Data validation (included with FastAPI)
]

[project.optional-dependencies]
ml = ["detoxify~=0.5", "torch>=2.0,<3.0"]
dev = [
    "httpx~=0.27",  # required by fastapi.testclient
    "pytest-asyncio",
    "pre-commit",  # git hooks for code quality
    "websockets",  # WebSocket client for testing
    "detoxify~=0.5",  # ML moderation
    "torch>=2.0,<3.0",  # Required by detoxify
    "mypy",  # Type checking
]

dash = [
    "streamlit",
    "pandas",
    "watchdog"
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "C4", "UP"]
ignore = ["E501", "B008"]  # Line too long, handled by black; B008 false positive for FastAPI

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-q --tb=short" 